Implement hugulp next
The main concept for hugulp next is to have it process the output of hugo,
instead of driving hugo all the time.

The workflow now becomes:

- Create your site normally as per hugo's guides
- Run hugo to build the "production" site (stored in the public folder)
- Run hugulp build to optimize the output of hugo

This would work perfectly, if it wasn't for sass/less, that need
pre-processing.

So, we need to keep a hugulp watch task, to transforms sass/less.

In the spirit of being flexible, we use a config file (.hugulprc),
that allows to configure the location of different folders.

Additionally, the config file allows to define which tasks to run. If
you don't want to fingerprint the assets, just remove the taks from
the pipeline field.

At this point, some more polishing is needed and probably a couple of
design decision can be changed, but the overall concept is pretty much
set.